add: _find_fw_structure; update: _load_structure_from_dir In an effort to allow more complicated subclasses of Structure to be reconstructed from a Firework directory, I've written some code that attempts to load the proper subclass instance of Structure, based on the initial instance and its update_sites() method. This basically means that the initial instance of that Structure-subclass object will be loaded, and then the sites will be updated based on the output of the directory.
Note: This is mainly because we've previously designed subclasses of Structure that have empty sites (Cathode), or details on the surface miller indices etc (QSlab). Passing this information about these structures down to subsequent calculations might be important, e.g. when we try to figure out the correct dimers for a Cathode object.