fix: issues with finding the correct structure from the FW.json file.
It seems that extracting the structure from the Vasp input set is not possible, because e.g.:

MPRelaxSet.from_dict(MPRelaxSet(structure).as_dict())

returns the class MPRelaxSet! It seems there are some issues in the from_dict method when applied to the subclasses of DictSet. This will have to be fixed.

For now, we simply add the structure/slab to the WriteVaspFromIOTask, so they can be found without having to rely on the vasp_input_set key.