testing: ability to add list of ErrorHandlers to Firework
One of the issues I have with the atomate fireworks is that they all use Custodian by default, and that you cannot specify your own ErrorHandlers. This raises quite a bit of problems when Custodian starts "correcting" error output which don't influence the results. Moreover, it doesn't allow me to specify my own corrections or post-calculations tests. I've implemented the option of passing a list of ErrorHandlers (i.e. a list of classes!) to a Firework, which will be used by the Custodian.