testing: ability to add list of ErrorHandlers to Firework
Because a class has no as_dict() method, I need to pass the ErrorHandlers as instances. This is a little less clean, as the output_filename of the handlers will have to be updated in the CustodianTask. It might be confusing for the user as well, as he or she might think that they need to provide the correct output file. Still, I can think of no better solution at this point.

One of the issues I have with the atomate fireworks is that they all use Custodian by default, and that you cannot specify your own ErrorHandlers. This raises quite a bit of problems when Custodian starts "correcting" error output which don't influence the results. Moreover, it doesn't allow me to specify my own corrections or post-calculations tests. I've implemented the option of passing a list of ErrorHandlers to a Firework, which will be used by the Custodian.