testing: more conditions for PulayTask
So far, PulayTask only checked the change in the lattice vectors to determine whether or not another geometry step should be executed. I've now added two other conditions: energy and ionic_steps (new default), each with their own tolerances. This way, the user can specify what the condition should be for performing another pulay step.

Now testing the new code on the cluster. Once it is working as intended, I'll add a more detailed description of the conditions.