fix: only define defaults for pulay step conditions in PulayTask
Currently, there are two locations where the 'default' condition for executing another geometry optimization is defined, i.e. in the PulayTask and the OptimizeFW. This is confusing and can lead to errors. Now the default is only defined at the PulayTask, but the code is a little less transparent because of it. There might be a more elegant solution for dealing with the fact that we want to pass the pulay conditions defined by default as None for the OptimizeFW to the PulayTask.