add: vasp_input_update
Added quick utility function that updates a dictionary that details the VASP input parameters to be passed to the DictSet (sub)class. Simply updating doesn't work because e.g. the user_incar_settings dictionary should usually not be entirely overwritten by the updated setting in case the user specifies a single INCAR tag. However, the user_kpoints_settings _should_ be overwritten entirely, as in case the Firework has a default user_kpoints_setting - which the user wants to overwrite with vasp_input_params - adding two keywords that define the kpoint generation scheme creates issues. This is due to the fact that DictSet goes down a list of options and simply chooses the first keyword that it finds in the dictionary. In other words, it's possible that the user chosen setting is ignored in case the default setting is first in line.