add: low CSHIFT to OpticsFW default INCAR params
For the SLME calculations, having a low CSHIFT is necessary to avoid smearing the imaginary part of the dielectric tensor significantly, which leads to an earlier onset in the absorptivity. VASP 5.4.4 has a strange way of deciding on whether to overwrite the original imaginary dielectric tensor by the smeared one from the Kramers-Kronig relation: if the resolution of the energy mesh (determined by the number of energy grid points NEDOS and the range of the energy interval) is smaller than the CSHIFT value, the imaginary part is smeared.

A value of CSHIFT=0.01 usually avoids smearing the dielectric tensor in the Kramers-Kronig integration. However, it is possible that for a large NEDOS CSHIFT is not sufficiently small. Moreover, the low CSHIFT value might result in issues in the Kramers-Kronig integration when determining the real part. It might be a good idea to design a check for this using a Custodian Errorhandler.