add: VaspParallelizationTask optimal_ncore optional argument
The VaspParallelizationTask sets the npar value based on the optimal ncore, which needs to be a divisor of the number of cores per kpoint. Currently, the optimal value for ncore is taken to be the one closes to the sqrt of the number of cores in a node. However, further testing indicates that for several architectures, 8 is a good choice for NCORE. We've set this as the default value which the divisor has to be close to, but allow for the user to set this argument in case they've found that another value works better for their architecture.

Note: This might actually be best set in the fworker environment... -> TODO