add: ParallelizationTestMonitor
Added a monitor ErrorHandler for parallelization tests. Parallelization test runs don't need to run for more than a specified number of electronic steps (enough to take a reasonable average), and should stop in case the time for any step is above a specified threshold (indicating that the setting is not good).