add: get_wf_parallel
First draft of the get_wf_parallel method. This workflow is used to find the optimal parallelization settings for a certain system, usually performed in the context of applying for computational resources on higher Tier machines. It is recommended to supply the workflows with suitable handlers that deal with issues due to the parallelization (JobTerminator) and make sure that not too much computational resources are used (ParallelizationTestMonitor), as some parallelization settings cause jobs to crash or run very slowly.