add: user_incar_settings argument to get_wf_optics()
It's tedious that I can't specify the user_incar_settings for the workflows, so I've added an argument for doing this (at least for two workflows so far). However, it's also not ideal that now the user_incar_settings are applied to ALL fireworks in the workflow. Maybe a list of dicts would make more sense, but I'm not sure how user friendly that is...

Perhaps in case the user wants to customize the workflow to such an extent, it makes more sense that they simply build it themselves from the Fireworks?