add: first version of optics module
This module has been largely copied from the module I wrote for the JARVIS package. Installing this entire package just because I need the classes I defined for post-processing optical calculations is a little tedious. It's probably better to add this module to pymatgen or ASE though, as it doesn't really belong in vsc-workflows.