add: POTCAR_FUNCTIONAL to calculation sets
Since commit b4b88d75ba5846e5542b1bb027a61521730d69d5, the preferred POTCAR set (e.g. PBE_54) can/must be specified in the input set. In case it is not specified, warnings are raised, and I run into issues when trying to run the workflows. -> Added the 'PBE' as default POTCAR_FUNCTIONAL for both the optimizeSet and staticSet.